+++ src/server/services/agentRuntime/__tests__/HumanInterventionHandler.test.ts
@@ -1,42 +1,20 @@
 // @vitest-environment node
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { AgentRuntimeService } from '../AgentRuntimeService';
+import { HumanInterventionHandler } from '../HumanInterventionHandler';
 
-// Mock heavy dependencies
-vi.mock('@/envs/app', () => ({ appEnv: { APP_URL: 'http://localhost:3010' } }));
-vi.mock('@/database/models/message', () => ({
-  MessageModel: vi.fn().mockImplementation(() => ({
-    updateMessagePlugin: vi.fn().mockResolvedValue(undefined),
-    updateToolMessage: vi.fn().mockResolvedValue({ success: true }),
-  })),
-}));
-vi.mock('@/server/modules/AgentRuntime', () => ({
-  AgentRuntimeCoordinator: vi.fn().mockImplementation(() => ({})),
-  createStreamEventManager: vi.fn(() => ({})),
-}));
-vi.mock('@/server/modules/AgentRuntime/RuntimeExecutors', () => ({
-  createRuntimeExecutors: vi.fn(() => ({})),
-}));
-vi.mock('@/server/services/mcp', () => ({ mcpService: {} }));
-vi.mock('@/server/services/queue', () => ({
-  QueueService: vi.fn().mockImplementation(() => ({})),
-}));
-vi.mock('@/server/services/queue/impls', () => ({ LocalQueueServiceImpl: class {} }));
-vi.mock('@/server/services/toolExecution', () => ({
-  ToolExecutionService: vi.fn().mockImplementation(() => ({})),
-}));
-vi.mock('@/server/services/toolExecution/builtin', () => ({
-  BuiltinToolsExecutor: vi.fn().mockImplementation(() => ({})),
-}));
-vi.mock('@lobechat/builtin-tools/dynamicInterventionAudits', () => ({
-  dynamicInterventionAudits: [],
-}));
+const buildHandler = (
+  pluginQuery: ReturnType<typeof vi.fn>,
+  messageModel: { updateMessagePlugin: any; updateToolMessage: any },
+) => {
+  const serverDB = { query: { messagePlugins: { findFirst: pluginQuery } } } as any;
+  return new HumanInterventionHandler(serverDB, messageModel as any);
+};
 
-describe('AgentRuntimeService.handleHumanIntervention', () => {
-  let service: AgentRuntimeService;
-  let mockMessageModel: any;
-  let mockDBPluginQuery: any;
+describe('HumanInterventionHandler.process', () => {
+  let mockMessageModel: { updateMessagePlugin: any; updateToolMessage: any };
+  let mockDBPluginQuery: ReturnType<typeof vi.fn>;
+  let handler: HumanInterventionHandler;
 
   const makeState = (overrides: Record<string, any> = {}) => ({
     lastModified: new Date().toISOString(),
@@ -50,25 +28,19 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
 
   beforeEach(() => {
     vi.clearAllMocks();
-
     mockDBPluginQuery = vi.fn().mockResolvedValue({ toolCallId: 'tool-call-1' });
-    const serverDB = {
-      query: { messagePlugins: { findFirst: mockDBPluginQuery } },
-    } as any;
-
-    service = new AgentRuntimeService(serverDB, 'user-1', { queueService: null });
     mockMessageModel = {
       updateMessagePlugin: vi.fn().mockResolvedValue(undefined),
       updateToolMessage: vi.fn().mockResolvedValue({ success: true }),
     };
-    (service as any).messageModel = mockMessageModel;
+    handler = buildHandler(mockDBPluginQuery, mockMessageModel);
   });
 
   describe('approve path', () => {
     it('persists intervention=approved on the tool message', async () => {
       const state = makeState();
 
-      await (service as any).handleHumanIntervention({} as any, state, {
+      await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
         toolMessageId: 'tool-msg-1',
       });
@@ -81,7 +53,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('returns nextContext with phase=human_approved_tool and skipCreateToolMessage=true', async () => {
       const state = makeState();
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
         toolMessageId: 'tool-msg-1',
       });
@@ -99,7 +71,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('removes the approved tool from pendingToolsCalling', async () => {
       const state = makeState();
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
         toolMessageId: 'tool-msg-1',
       });
@@ -111,7 +83,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('keeps state waiting_for_human while other tools still pending', async () => {
       const state = makeState();
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
         toolMessageId: 'tool-msg-1',
       });
@@ -126,7 +98,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
         ],
       });
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
         toolMessageId: 'tool-msg-1',
       });
@@ -137,7 +109,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('no-ops when toolMessageId is missing', async () => {
       const state = makeState();
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
       });
 
@@ -150,7 +122,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('persists intervention=rejected with reason and updates content', async () => {
       const state = makeState();
 
-      await (service as any).handleHumanIntervention({} as any, state, {
+      await handler.process(state, {
         rejectionReason: 'privacy concern',
         toolMessageId: 'tool-msg-1',
       });
@@ -163,30 +135,21 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
       });
     });
 
-    it('uses default content when no reason provided', async () => {
+    it('does not enter the reject branch when rejectionReason is falsy', async () => {
       const state = makeState();
 
-      await (service as any).handleHumanIntervention({} as any, state, {
+      await handler.process(state, {
         rejectionReason: '',
         toolMessageId: 'tool-msg-1',
       });
 
-      // Empty string is falsy so it won't enter the reject branch. Cover the
-      // "no reason" content path by passing a space-only reason explicitly:
-      // the branch is "reason ? withReason : withoutReason" inside the handler.
-      // We verify the with-reason branch above; the without-reason branch is
-      // covered below via an explicit sentinel.
       expect(mockMessageModel.updateToolMessage).not.toHaveBeenCalled();
     });
 
-    it('writes "without reason" content when reason is whitespace', async () => {
-      // handleHumanIntervention treats the rejection as present whenever
-      // rejectionReason is truthy, then chooses content based on truthiness
-      // of the trimmed reason. We pass a non-empty sentinel to ensure the
-      // branch runs but assert the literal "with reason" template by value.
+    it('writes "with reason" content for any non-empty reason', async () => {
       const state = makeState();
 
-      await (service as any).handleHumanIntervention({} as any, state, {
+      await handler.process(state, {
         rejectionReason: 'r',
         toolMessageId: 'tool-msg-1',
       });
@@ -203,7 +166,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
       const state = makeState();
       mockDBPluginQuery.mockResolvedValueOnce({ toolCallId: 'tool-call-2' });
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         rejectionReason: 'nope',
         toolMessageId: 'tool-msg-2',
       });
@@ -215,7 +178,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('transitions to interrupted + reason=human_rejected (pure reject, no continue)', async () => {
       const state = makeState();
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         rejectionReason: 'nope',
         toolMessageId: 'tool-msg-1',
       });
@@ -239,7 +202,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
       const state = makeState();
       mockDBPluginQuery.mockResolvedValueOnce({ toolCallId: 'tool-call-1' });
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         rejectAndContinue: true,
         rejectionReason: 'nope',
         toolMessageId: 'tool-msg-1',
@@ -257,7 +220,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
       });
       mockDBPluginQuery.mockResolvedValueOnce({ toolCallId: 'tool-call-1' });
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         rejectAndContinue: true,
         rejectionReason: 'nope',
         toolMessageId: 'tool-msg-1',
@@ -270,7 +233,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('still persists intervention=rejected on the tool message', async () => {
       const state = makeState();
 
-      await (service as any).handleHumanIntervention({} as any, state, {
+      await handler.process(state, {
         rejectAndContinue: true,
         rejectionReason: 'privacy',
         toolMessageId: 'tool-msg-1',
@@ -286,7 +249,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('returns state unchanged when status is not waiting_for_human (approve)', async () => {
       const state = makeState({ status: 'running' });
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         approvedToolCall: { id: 'tool-call-1' },
         toolMessageId: 'tool-msg-1',
       });
@@ -299,7 +262,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('returns state unchanged when status is not waiting_for_human (reject)', async () => {
       const state = makeState({ status: 'running' });
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         rejectionReason: 'nope',
         toolMessageId: 'tool-msg-1',
       });
@@ -311,7 +274,7 @@ describe('AgentRuntimeService.handleHumanIntervention', () => {
     it('handles humanInput as out-of-scope (no state transition)', async () => {
       const state = makeState();
 
-      const result = await (service as any).handleHumanIntervention({} as any, state, {
+      const result = await handler.process(state, {
         humanInput: { response: 'hi' },
         toolMessageId: 'tool-msg-1',
       });
